EMO-7023 blob storage refactoring
