+++ blob/src/main/java/com/bazaarvoice/emodb/blob/core/DefaultBlobStore.java
@@ -10,12 +10,12 @@ import com.bazaarvoice.emodb.blob.api.DefaultTable;
 import com.bazaarvoice.emodb.blob.api.Names;
 import com.bazaarvoice.emodb.blob.api.Range;
 import com.bazaarvoice.emodb.blob.api.RangeSpecification;
-import com.bazaarvoice.emodb.blob.api.StreamSupplier;
 import com.bazaarvoice.emodb.blob.db.MetadataProvider;
 import com.bazaarvoice.emodb.blob.db.StorageProvider;
 import com.bazaarvoice.emodb.blob.db.StorageSummary;
 import com.bazaarvoice.emodb.common.api.impl.LimitCounter;
 import com.bazaarvoice.emodb.sor.api.Audit;
+import com.bazaarvoice.emodb.sor.api.TableAvailability;
 import com.bazaarvoice.emodb.sor.api.TableExistsException;
 import com.bazaarvoice.emodb.sor.api.TableOptions;
 import com.bazaarvoice.emodb.sor.api.UnknownTableException;
@@ -23,35 +23,29 @@ import com.bazaarvoice.emodb.table.db.Table;
 import com.bazaarvoice.emodb.table.db.TableDAO;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
-import com.google.common.base.Function;
 import com.google.common.base.Throwables;
-import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Maps;
-import com.google.common.io.ByteStreams;
 import com.google.common.io.InputSupplier;
 import com.google.inject.Inject;
-import org.apache.commons.codec.binary.Hex;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
-import java.security.DigestInputStream;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.Collection;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.Map;
+import java.util.Spliterator;
+import java.util.Spliterators;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static java.lang.String.format;
 
 /**
  * Stores large binary objects like photos, videos.
@@ -88,21 +82,16 @@ public class DefaultBlobStore implements BlobStore {
     @Override
     public Iterator<com.bazaarvoice.emodb.blob.api.Table> listTables(@Nullable String fromTableExclusive, long limit) {
         checkArgument(limit > 0, "Limit must be >0");
-
         LimitCounter remaining = new LimitCounter(limit);
-        final Iterator<Table> tableIter = _tableDao.list(fromTableExclusive, remaining);
-        return remaining.limit(new AbstractIterator<com.bazaarvoice.emodb.blob.api.Table>() {
-            @Override
-            protected com.bazaarvoice.emodb.blob.api.Table computeNext() {
-                while (tableIter.hasNext()) {
-                    Table table = tableIter.next();
-                    if (!table.isInternal()) {
-                        return toDefaultTable(table);
-                    }
-                }
-                return endOfData();
-            }
-        });
+        return getTablesStream(fromTableExclusive, remaining)
+                .map(DefaultBlobStore::toDefaultTable)
+                .iterator();
+    }
+
+    private Stream<Table> getTablesStream(@Nullable String fromTableExclusive, LimitCounter remaining) {
+        return StreamSupport.stream(Spliterators.spliteratorUnknownSize(_tableDao.list(fromTableExclusive, remaining), Spliterator.ORDERED), false)
+                .filter(table -> !table.isInternal())
+                .limit(remaining.remaining());
     }
 
     private static com.bazaarvoice.emodb.blob.api.Table toDefaultTable(Table table) {
@@ -139,7 +128,11 @@ public class DefaultBlobStore implements BlobStore {
     public void purgeTableUnsafe(String tableName, Audit audit) throws UnknownTableException {
         checkLegalTableName(tableName);
         Table table = _tableDao.get(tableName);
-        _tableDao.audit(tableName, "purge", audit);
+        purgeTableUnsafe(table, audit);
+    }
+
+    private void purgeTableUnsafe(Table table, Audit audit) {
+        _tableDao.audit(table.getName(), "purge", audit);
 
         AtomicLong failedCounter = new AtomicLong();
         AtomicLong totalCounter = new AtomicLong();
@@ -154,7 +147,7 @@ public class DefaultBlobStore implements BlobStore {
             }
         });
 
-        if(totalCounter.get() > 0) {
+        if (totalCounter.get() > 0) {
             if (failedCounter.get() > 0) {
                 String message = String.format("Failed to purge %s of %s rows for table: %s.", failedCounter.get(), totalCounter.get(), table.getName());
                 LOGGER.error(message);
@@ -165,7 +158,6 @@ public class DefaultBlobStore implements BlobStore {
         } else {
             LOGGER.info("Attempting to purge table: {} that is already empty.", table.getName());
         }
-
     }
 
     @Override
@@ -235,12 +227,7 @@ public class DefaultBlobStore implements BlobStore {
         // Stream back results.  Don't hold them all in memory at once.
         LimitCounter remaining = new LimitCounter(limit);
         return remaining.limit(Iterators.transform(_metadataProvider.scanMetadata(table, fromBlobIdExclusive, remaining),
-                new Function<Map.Entry<String, StorageSummary>, BlobMetadata>() {
-                    @Override
-                    public BlobMetadata apply(Map.Entry<String, StorageSummary> entry) {
-                        return newMetadata(table, entry.getKey(), entry.getValue());
-                    }
-                }));
+                entry -> newMetadata(table, entry.getKey(), entry.getValue())));
     }
 
     private static BlobMetadata newMetadata(Table table, String blobId, StorageSummary s) {
@@ -283,68 +270,7 @@ public class DefaultBlobStore implements BlobStore {
             range = new Range(0, summary.getLength());
         }
 
-        return new DefaultBlob(metadata, range, new StreamSupplier() {
-            @Override
-            public void writeTo(OutputStream out) throws IOException {
-                readChunks(table, blobId, range, summary, out);
-            }
-        });
-    }
-
-    private void readChunks(Table table, String blobId, Range range, StorageSummary summary, OutputStream out)
-            throws IOException {
-        if (range.getLength() == 0) {
-            return; // Nothing to do
-        }
-        // Calculate the location of the first byte
-        long start = range.getOffset();
-        int startChunk = (int) (start / summary.getChunkSize());
-        int startOffset = (int) (start % summary.getChunkSize());
-        // Calculate the location of the last byte
-        long end = range.getOffset() + range.getLength() - 1; // Inclusive
-        int endChunk = (int) (end / summary.getChunkSize());
-        int endLimit = (int) (end % summary.getChunkSize()) + 1; // Exclusive
-
-        // A more aggressive solution like the Astyanax ObjectReader recipe would improve performance by issuing
-        // the reads in parallel.
-        for (int i = startChunk; i <= endChunk; i++) {
-            ByteBuffer chunk = _storageProvider.readChunk(table, blobId, i, summary.getTimestamp());
-            if (chunk == null) {
-                throw new IOException(format("Blob chunk %d is missing: %s", i, blobId));
-            }
-
-            // Adjust the start and end of the byte buffer if fetching a range of bytes, not the entire blob.
-            int position = chunk.position();
-            if (i == startChunk) {
-                chunk.position(position + startOffset);
-            }
-            if (i == endChunk) {
-                chunk.limit(position + endLimit);
-            }
-
-            // Copy the chunk bytes to the output stream.
-            copyTo(chunk, out);
-        }
-    }
-
-    /**
-     * Copy the contents of a ByteBuffer to an OutputStream.
-     */
-    private static void copyTo(ByteBuffer buf, OutputStream out) throws IOException {
-        if (!buf.hasRemaining()) {
-            return;
-        }
-        if (buf.hasArray()) {
-            // Fast copy if the buffer is backed by an array (which should be the case)
-            out.write(buf.array(), buf.arrayOffset() + buf.position(), buf.remaining());
-        } else {
-            // Slow copy otherwise
-            byte[] bytes = new byte[4096];
-            do {
-                buf.get(bytes, 0, Math.min(bytes.length, buf.remaining()));
-                out.write(bytes);
-            } while (buf.hasRemaining());
-        }
+        return new DefaultBlob(metadata, range, _storageProvider.getObjectStreamSupplier(table, blobId, summary, range));
     }
 
     @Override
@@ -356,7 +282,7 @@ public class DefaultBlobStore implements BlobStore {
 
         Table table = _tableDao.get(tableName);
 
-        StorageSummary summary = putObject(table, blobId, in, attributes);
+        StorageSummary summary = _storageProvider.putObject(table, blobId, in.getInput(), attributes);
 
         try {
             _metadataProvider.writeMetadata(table, blobId, summary);
@@ -372,42 +298,6 @@ public class DefaultBlobStore implements BlobStore {
                 Throwables.propagate(t);
             }
         }
-    }
-
-    private StorageSummary putObject(Table table, String blobId, InputSupplier<? extends InputStream> in, Map<String, String> attributes) throws IOException {
-        long timestamp = _storageProvider.getCurrentTimestamp(table);
-        int chunkSize = _storageProvider.getDefaultChunkSize();
-        checkArgument(chunkSize > 0);
-        DigestInputStream md5In = new DigestInputStream(in.getInput(), getMessageDigest("MD5"));
-        DigestInputStream sha1In = new DigestInputStream(md5In, getMessageDigest("SHA-1"));
-
-        // A more aggressive solution like the Astyanax ObjectWriter recipe would improve performance by pipelining
-        // reading the input stream and writing chunks, and issuing the writes in parallel.
-        byte[] bytes = new byte[chunkSize];
-        long length = 0;
-        int chunkCount = 0;
-        for (; ; ) {
-            int chunkLength;
-            try {
-                chunkLength = ByteStreams.read(sha1In, bytes, 0, bytes.length);
-            } catch (IOException e) {
-                LOGGER.error("Failed to read input stream", e);
-                throw Throwables.propagate(e);
-            }
-            if (chunkLength == 0) {
-                break;
-            }
-            ByteBuffer buffer = ByteBuffer.wrap(bytes, 0, chunkLength);
-            _storageProvider.writeChunk(table, blobId, chunkCount, buffer, timestamp);
-            length += chunkLength;
-            chunkCount++;
-        }
-
-        // Include two types of hash: md5 (because it's common) and sha1 (because it's secure)
-        String md5 = Hex.encodeHexString(md5In.getMessageDigest().digest());
-        String sha1 = Hex.encodeHexString(sha1In.getMessageDigest().digest());
-
-        return new StorageSummary(length, chunkCount, chunkSize, md5, sha1, attributes, timestamp);
     }
 
     @Override
@@ -454,14 +344,6 @@ public class DefaultBlobStore implements BlobStore {
         return (Map) table.getAttributes();
     }
 
-    private static MessageDigest getMessageDigest(String algorithmName) {
-        try {
-            return MessageDigest.getInstance(algorithmName);
-        } catch (NoSuchAlgorithmException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
     private static void checkLegalTableName(String table) {
         checkArgument(Names.isLegalTableName(table),
                 "Table name must be a lowercase ASCII string between 1 and 255 characters in length. " +
@@ -474,4 +356,5 @@ public class DefaultBlobStore implements BlobStore {
                 "Blob IDs must be ASCII strings between 1 and 255 characters in length. " +
                         "Whitespace, ISO control characters and certain punctuation characters that aren't generally allowed in file names are excluded.");
     }
+
 }
